requirements.txt
asgiref==3.5.0
Django==4.0.2
django-admin-thumbnails==0.2.6
django-tinymce==3.4.0
jedi==0.18.1
parso==0.8.3
Pillow==9.0.1
regex==2022.1.18
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
